requirements.txt
requests>=2.28
openai>=1.0
pyyaml>=6.0
